♻️ separate conversation stream contract
