+++ eslint.config.mjs
@@ -132,14 +132,14 @@ export default [
             { sourceTag: "scope:agent-runtime-controller", onlyDependOnLibsWithTags: ["scope:agent-runtime-controller", "scope:agent-runtime-launcher", "scope:shared"] },
             { sourceTag: "scope:agent-controller", onlyDependOnLibsWithTags: ["scope:agent-controller", "scope:agent-runtime-controller", "scope:skills-controller", "scope:shared"] },
             { sourceTag: "scope:cluster-tenants", onlyDependOnLibsWithTags: ["scope:auth", "scope:cluster-tenants", "scope:k8s-api", "scope:shared"] },
-			{ sourceTag: "scope:conversation-projection", onlyDependOnLibsWithTags: ["scope:conversation-projection", "scope:shared"] },
+			{ sourceTag: "scope:conversation-projection", onlyDependOnLibsWithTags: ["scope:conversation-agent-threads", "scope:conversation-projection", "scope:shared"] },
 			{ sourceTag: "scope:conversation-agent-threads", onlyDependOnLibsWithTags: ["scope:conversation-agent-threads", "scope:shared"] },
 			{ sourceTag: "scope:conversations", onlyDependOnLibsWithTags: ["scope:agents", "scope:auth", "scope:channel-targets", "scope:conversation-agent-threads", "scope:conversation-projection", "scope:conversations", "scope:execution-admission", "scope:execution-runs", "scope:shared"] },
 			{ sourceTag: "scope:conversation-assets", onlyDependOnLibsWithTags: ["scope:artifacts", "scope:auth", "scope:conversations", "scope:conversation-assets", "scope:execution-runs", "scope:shared", "scope:web"] },
 			{ sourceTag: "scope:personal-configuration", onlyDependOnLibsWithTags: ["scope:agent-services", "scope:agents", "scope:auth", "scope:personal-configuration", "scope:shared"] },
 			{ sourceTag: "scope:user-onboarding", onlyDependOnLibsWithTags: ["scope:user-onboarding", "scope:shared"] },
             { sourceTag: "scope:grants", onlyDependOnLibsWithTags: ["scope:auth", "scope:authorization", "scope:grants", "scope:shared"] },
-            { sourceTag: "scope:groups", onlyDependOnLibsWithTags: ["scope:groups", "scope:shared"] },
+            { sourceTag: "scope:groups", onlyDependOnLibsWithTags: ["scope:auth", "scope:groups", "scope:shared"] },
             { sourceTag: "scope:http", onlyDependOnLibsWithTags: ["scope:http", "scope:shared"] },
             { sourceTag: "scope:integrations", onlyDependOnLibsWithTags: ["scope:agents", "scope:auth", "scope:integrations", "scope:obot-custody", "scope:shared"] },
             { sourceTag: "scope:channel-targets", onlyDependOnLibsWithTags: ["scope:auth", "scope:authorization", "scope:channel-targets", "scope:membership", "scope:shared"] },
@@ -175,7 +175,7 @@ export default [
             { sourceTag: "scope:skills", onlyDependOnLibsWithTags: ["scope:artifacts", "scope:auth", "scope:cluster-tenants", "scope:grants", "scope:shared", "scope:skills"] },
 			{ sourceTag: "scope:spend", onlyDependOnLibsWithTags: ["scope:shared", "scope:spend"] },
 			{ sourceTag: "scope:conversation-elicitation", onlyDependOnLibsWithTags: ["scope:conversation-elicitation", "scope:shared", "scope:web"] },
-			{ sourceTag: "scope:agent-threads", onlyDependOnLibsWithTags: ["scope:agent-threads", "scope:conversation-assets", "scope:conversation-elicitation", "scope:shared", "scope:web"] },
+			{ sourceTag: "scope:agent-threads", onlyDependOnLibsWithTags: ["scope:agent-threads", "scope:conversation-assets", "scope:conversation-elicitation", "scope:conversations", "scope:shared", "scope:web"] },
 			{ sourceTag: "scope:conversation-workspace", onlyDependOnLibsWithTags: ["scope:agent-threads", "scope:conversation-assets", "scope:conversation-elicitation", "scope:conversation-workspace", "scope:conversations", "scope:shared", "scope:user-onboarding", "scope:web"] },
 			{ sourceTag: "scope:web", onlyDependOnLibsWithTags: ["scope:web", "scope:shared"] },
 			{ sourceTag: "scope:opencrane-ui", onlyDependOnLibsWithTags: ["scope:agent-threads", "scope:conversation-assets", "scope:conversation-workspace", "scope:persona-onboarding", "scope:shared", "scope:web"] },
@@ -227,7 +227,7 @@ export default [
             { sourceTag: "type:app", notDependOnLibsWithTags: ["type:app"] },
 			{ sourceTag: "type:lib", notDependOnLibsWithTags: ["type:app"] },
 			{ sourceTag: "frontend-role:feature", onlyDependOnLibsWithTags: ["frontend-role:elements", "frontend-role:state"] },
-			{ sourceTag: "frontend-role:feature-shell", onlyDependOnLibsWithTags: ["frontend-role:elements", "frontend-role:elements-composite", "frontend-role:feature", "frontend-role:state", "frontend-role:state-composite"] },
+			{ sourceTag: "frontend-role:feature-shell", onlyDependOnLibsWithTags: ["frontend-role:core", "frontend-role:elements", "frontend-role:elements-composite", "frontend-role:feature", "frontend-role:state", "frontend-role:state-composite", "layer:contract", "layer:model", "type:platform"] },
 			{ sourceTag: "frontend-role:adapter", onlyDependOnLibsWithTags: ["frontend-role:core", "frontend-role:state", "frontend-role:state-composite", "layer:contract", "layer:model"] },
 			{ sourceTag: "frontend-role:state", onlyDependOnLibsWithTags: ["frontend-role:core", "layer:contract", "layer:model", "layer:util"] },
 			{ sourceTag: "frontend-role:state-composite", onlyDependOnLibsWithTags: ["frontend-role:core", "frontend-role:state", "layer:contract", "layer:model", "layer:util"] },
