+++ libs/frontend/state/conversation/adapter/src/index.ts
@@ -1,2 +1,1 @@
-export { ConversationEventStreamStatuses, type ConversationEventStream, type ConversationEventStreamUpdate, type StreamConversationEventsCommand } from "./lib/conversation-event-stream.types";
 export { OpenCraneConversationEventStream } from "./lib/opencrane-conversation-event-stream";
