+++ libs/frontend/state/conversation/ag-ui/project.json
@@ -3,7 +3,7 @@
   "$schema": "../../../../../node_modules/nx/schemas/project-schema.json",
   "projectType": "library",
   "sourceRoot": "libs/frontend/state/conversation/ag-ui/src",
-  "tags": ["scope:web", "type:state", "frontend-role:state"],
+  "tags": ["scope:web", "type:state", "frontend-role:core"],
   "targets": {
     "lint": { "executor": "nx:run-commands", "options": { "command": "tsc -p tsconfig.lint.json", "cwd": "libs/frontend/state/conversation/ag-ui" } },
     "test": { "executor": "nx:run-commands", "options": { "command": "vitest run --passWithNoTests", "cwd": "libs/frontend/state/conversation/ag-ui" } }
