+++ libs/frontend/state/conversation/stream/project.json
@@ -1,0 +1,11 @@
+{
+  "name": "frontend-conversation-stream-state",
+  "$schema": "../../../../../node_modules/nx/schemas/project-schema.json",
+  "projectType": "library",
+  "sourceRoot": "libs/frontend/state/conversation/stream/src",
+  "tags": ["scope:web", "type:lib", "layer:frontend", "frontend-role:state"],
+  "targets": {
+    "lint": { "executor": "nx:run-commands", "options": { "command": "tsc -p tsconfig.lint.json", "cwd": "libs/frontend/state/conversation/stream" } },
+    "test": { "executor": "nx:run-commands", "options": { "command": "vitest run --passWithNoTests", "cwd": "libs/frontend/state/conversation/stream" } }
+  }
+}
