+++ libs/frontend/state/conversation/stream/tsconfig.json
@@ -1,0 +1,1 @@
+{ "extends": "../../../../../tsconfig.frontend.json", "compilerOptions": { "strict": true }, "files": [], "include": [], "references": [{ "path": "./tsconfig.lib.json" }] }
