+++ libs/frontend/state/conversation/stream/tsconfig.lib.json
@@ -1,0 +1,1 @@
+{ "extends": "./tsconfig.json", "compilerOptions": { "outDir": "../../../../../dist/out-tsc", "types": [], "composite": true, "tsBuildInfoFile": "../../../../../dist/out-tsc/libs/frontend/state/conversation/stream/tsconfig.lib.tsbuildinfo" }, "include": ["src/**/*.ts"], "exclude": ["src/**/*.spec.ts"], "references": [{ "path": "../ag-ui/tsconfig.lib.json" }] }
