+++ libs/frontend/state/conversation/stream/tsconfig.lint.json
@@ -1,0 +1,1 @@
+{ "extends": "./tsconfig.json", "compilerOptions": { "composite": false, "declaration": false, "incremental": false, "noEmit": true }, "include": ["src/**/*.ts"], "exclude": ["**/*.spec.ts"], "references": [] }
